created geUser(all and by id)/postUser(only uniq users)

+++ rest-api/api/models/posts.go
@@ -4,11 +4,11 @@ import "time"
 
 //Post is an...
 type Post struct {
-	ID       uint32 `gorm:"primary_key;auto_increment" json:"id"`
+	ID       uint64 `gorm:"primary_key;auto_increment" json:"id"`
 	Title    string `gorm:"type:varchar(255)" json:"title"`
 	ImageURL string `gorm:"type:varchar(255)" json:"image_url"`
 	Body     string `gorm:"type:text" json:"body"`
-	UserID   uint32 `gorm:"not null" json:"user_id"`
+	UserID   uint64 `gorm:"not null" json:"user_id"`
 
 	CreatedAt time.Time `gorm:"type:timestamp(0);default:current_timestamp();not null" json:"created_at"`
 	UpdatedAt time.Time `gorm:"type:timestamp(0);default:current_timestamp();not null" json:"updated_at"`
